+++ marioparty/scripts/icon.js
@@ -30,11 +30,11 @@ const rosalinaIcon = new Icon("Rosalina", "../images/characters/rosalina_icon.pn
 const lumaIcon = new Icon("Luma", "../images/characters/luma_icon.png", "#ffd700");
 const shyguyIcon = new Icon("Shy Guy", "../images/characters/shyguy_icon.png", "#aa0055");
 
-export const allCharacterIcons = [
+export let allCharacterIcons = [
 	marioIcon, luigiIcon, peachIcon, booIcon, paulineIcon, yoshiIcon, plantIcon, royIcon, waluigiIcon, bowserIcon,
 	rosalinaIcon, lumaIcon, shyguyIcon
 ];
-export const someCharacterIcons = [
+export let someCharacterIcons = [
 	marioIcon, luigiIcon, peachIcon, waluigiIcon, bowserIcon, shyguyIcon, rosalinaIcon
 ];
 
@@ -51,7 +51,7 @@ const paydayDoubleIcon = new Icon("Double Payday Dice", "../images/items/payday_
 const paydayTripleIcon = new Icon("Triple Payday Dice", "../images/items/payday_triple.png", null);
 const superMirrorIcon = new Icon("Super Mirror", "../images/items/super_mirror.png", null);
 
-export const itemSpaceIcons = [
+export let itemSpaceIcons = [
 	mushroomIcon, doubleDiceIcon, mirrorIcon, tripleDiceIcon, chompCallIcon, pipeIcon
 ];
 
@@ -66,26 +66,56 @@ const rightstar = new Icon("has to give a star to", "../images/chancetime/give_r
 const swapstars = new Icon("has to swap stars with", "../images/chancetime/swap_stars.png", "#afc4a0")
 const swapall = new Icon("has to swap both stars and coins with", "../images/chancetime/swap_all.png", "#aabc95")
 
-export const chanceTimeIcons = [
+export let chanceTimeIcons = [
 	left20, right20, left30, right30, swapcoins, leftstar, rightstar, swapstars, swapall
 ];
 
 // Game Icons
-const musicalChairsIcon = new Icon("Musical Chairs", "../images/other/question_mark.png", null); 
-const redlLightGreenLightIcon = new Icon("Red Light Green Light", "../images/other/question_mark.png", null);
-const messyMemoryIcon = new Icon("Messy Memory", "../images/other/question_mark.png", null); 
-const lucyCupIcon = new Icon("Lucky Cup", "../images/other/question_mark.png", null); 
-const coinPongIcon = new Icon("Coin Pong", "../images/other/question_mark.png", null); 
-const chopstickRelayIcon = new Icon("Chopstick Relay", "../images/other/question_mark.png", null); 
-const bobbingForApplesIcon = new Icon("Bobbing for Apples", "../images/other/question_mark.png", null); 
-const makingFacesIcon = new Icon("Making Faces", "../images/other/question_mark.png", null); 
-const luckOfTheDiceIcon = new Icon("Luck of the Dice", "../images/other/question_mark.png", null); 
-const passTheWaterIcon = new Icon("Pass the Water", "../images/other/question_mark.png", null); 
+const musicalChairsIcon = new Icon("Musical Chairs", "../images/other/question_mark.jpg", null); 
+const redlLightGreenLightIcon = new Icon("Red Light Green Light", "../images/other/question_mark.jpg", null);
+const messyMemoryIcon = new Icon("Messy Memory", "../images/other/question_mark.jpg", null); 
+const lucyCupIcon = new Icon("Lucky Cup", "../images/other/question_mark.jpg", null); 
+const coinPongIcon = new Icon("Coin Pong", "../images/other/question_mark.jpg", null); 
+const chopstickRelayIcon = new Icon("Chopstick Relay", "../images/other/question_mark.jpg", null); 
+const bobbingForApplesIcon = new Icon("Bobbing for Apples", "../images/other/question_mark.jpg", null); 
+const makingFacesIcon = new Icon("Making Faces", "../images/other/question_mark.jpg", null); 
+const luckOfTheDiceIcon = new Icon("Luck of the Dice", "../images/other/question_mark.jpg", null); 
+const passTheWaterIcon = new Icon("Pass the Water", "../images/other/question_mark.jpg", null); 
     
-export const freeForAllMinigameIcons = [
+export let freeForAllMinigameIcons = [
 	musicalChairsIcon, redlLightGreenLightIcon, messyMemoryIcon, lucyCupIcon, coinPongIcon
 ];
-export const teamMinigameIcons = [
+export let teamMinigameIcons = [
 	chopstickRelayIcon, bobbingForApplesIcon, makingFacesIcon, luckOfTheDiceIcon, passTheWaterIcon
 ];
 
+// Lucky Space Icons
+const recieve5CoinsIcon = new Icon("Recieve 5 Coins", "../images/items/coin.png", null);
+const recieve7CoinsIcon = new Icon("Recieve 7 Coins", "../images/items/coin.png", null);
+const recieve10CoinsIcon = new Icon("Recieve 10 Coins", "../images/items/coin.png", null);
+const recieve12CoinsIcon = new Icon("Recieve 12 Coins", "../images/items/coin_stack.png", null);
+const recieve15CoinsIcon = new Icon("Recieve 15 Coins", "../images/items/coin_stack.png", null);
+const recieve20CoinsIcon = new Icon("Recieve 20 Coins", "../images/items/coin_stack.png", null);
+const recievedoubleIcon = new Icon("Recieve Double Dice", "../images/items/double_dice.png", null);
+const recieveTripleIcon = new Icon("Recieve Triple Dice", "../images/items/triple_dice.png", null);
+const recieveCustomIcon = new Icon("Recieve a Custom Dice", "../images/items/custom_dice.png", null)
+const recieveMushroomIcon = new Icon("Recieve a Mushroom", "../images/items/mushroom.png", null);
+const recieveMirrorIcon = new Icon("Recieve a Mirror", "../images/items/mirror.png", null);
+
+export let allLuckyIcons = [
+	recieve5CoinsIcon, recieve7CoinsIcon, recieve10CoinsIcon, recieve12CoinsIcon, recieve15CoinsIcon, recieve20CoinsIcon,
+	recievedoubleIcon, recieveTripleIcon, recieveCustomIcon, recieveMushroomIcon, recieveMirrorIcon
+];
+
+// Bowser Space Icons
+const lose10Coins = new Icon("Lose 10 Coins", "../images/other/question_mark.jpg", null);
+const lose15Coins = new Icon("Lose 15 Coins", "../images/other/question_mark.jpg", null);
+const lose20Coins = new Icon("Lose 20 Coins", "../images/other/question_mark.jpg", null);
+const lose30Coins = new Icon("Lose 30 Coins", "../images/other/question_mark.jpg", null);
+const loseHalfCoins = new Icon("Lose Half your Coins", "../images/other/question_mark.jpg", null);
+const allLose10Coins = new Icon("Everyone Lose 10 Coins", "../images/other/question_mark.jpg", null);
+const loseAStar = new Icon("Lose a Star", "../images/other/question_mark.jpg", null);
+
+export let allBowserIcons = [
+	lose10Coins, lose15Coins, lose20Coins, lose30Coins, loseHalfCoins, allLose10Coins, loseAStar
+];